requirements.txt
discord.py[voice]>=1.3.2
pyyaml
pynacl
youtube-dl>=2020.3.24
Pillow
html5lib
beautifulsoup4
requests
psutil